+++ src/main/java/fuzs/goldenagecombat/config/ServerConfig.java
@@ -1,5 +1,6 @@
 package fuzs.goldenagecombat.config;
 
+import com.google.common.collect.Lists;
 import fuzs.puzzleslib.config.AbstractConfig;
 import fuzs.puzzleslib.config.annotation.Config;
 import fuzs.puzzleslib.config.serialization.EntryCollectionBuilder;
@@ -7,7 +8,6 @@ import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.registries.ForgeRegistries;
-import org.apache.commons.compress.utils.Lists;
 
 import java.util.List;
 import java.util.Map;
@@ -38,20 +38,24 @@ public class ServerConfig extends AbstractConfig {
         VANILLA, COMBAT_UPDATE, LEGACY_COMBAT, CUSTOM
     }
 
+    public enum UpwardsKnockback {
+        NONE, OLD_COMBAT, COMBAT_TESTS
+    }
+
     public static class ClassicConfig extends AbstractConfig {
         @Config(name = "remove_attack_cooldown", description = "Completely remove the attack cooldown as if it never even existed in the first place.")
         public boolean removeCooldown = true;
         @Config(name = "food_mechanics", description = {"Choose the food mechanics to use:", "VANILLA will change nothing and use surplus saturation for very quick health regeneration.", "LEGACY_COMBAT will make health only regenerated every 4 seconds, while requiring 18 or more food points.", "CUSTOM will make health only regenerated every 3 seconds, which requires more than 6 food points.", "COMBAT_TEST will make health regenerated every 2 seconds, which requires more than 6 food points. Also food points will be directly consumed when healing."})
         public FoodMechanics foodMechanics = FoodMechanics.CUSTOM;
-        @Config(name = "weak_attacks_knock_back_player", description = "Player is knocked back by attacks which do not cause any damage, such as when hit by snowballs and eggs.")
+        @Config(name = "weak_attacks_knock_back_player", description = "Player is knocked back by attacks which do not cause any damage, such as when hit by snowballs, eggs, and fishing rod hooks.")
         public boolean weakPlayerKnockback = true;
         @Config(name = "critical_hits_while_sprinting", description = "Sprinting and attacking no longer interfere with each other, making critical hits possible at all times.")
         public boolean criticalHitsSprinting = true;
-        @Config(name = "fishing_rod_causes_knockback", description = "Fishing rod deals knockback upon hitting an entity.")
+        @Config(category = "fishing_rod", name = "cause_knockback", description = "Fishing rod deals knockback upon hitting an entity.")
         public boolean fishingRodKnockback = true;
-        @Config(name = "fishing_rod_launches_entities", description = "Entities reeled in using a fishing rod are slightly launched upwards.")
+        @Config(category = "fishing_rod", name = "launch_entities", description = "Entities reeled in using a fishing rod are slightly launched upwards.")
         public boolean fishingRodLaunch = true;
-        @Config(name = "fishing_rod_breaks_slower", description = "Hooking entities with a fishing rod causes only 3 damage points to the rod instead of 5.")
+        @Config(category = "fishing_rod", name = "slower_breaking", description = "Hooking entities with a fishing rod causes only 3 damage points to the rod instead of 5.")
         public boolean fishingRodSlowerBreaking = true;
         @Config(name = "boost_sharpness", description = "Boost sharpness enchantment to 1.25 damage points per level instead of just 0.5.")
         public boolean boostSharpness = false;
@@ -66,7 +70,7 @@ public class ServerConfig extends AbstractConfig {
         @Config(name = "interact_while_using", description = "Allow using the \"Attack\" button while the \"Use Item\" button is held for mining blocks. Does not make a lot of sense, but it used to be a feature in old pvp.")
         public boolean interactWhileUsing = false;
         @Config(name = "upwards_knockback", description = "Makes knockback stronger towards targets not on the ground.")
-        public boolean upwardsKnockback = true;
+        public UpwardsKnockback upwardsKnockback = UpwardsKnockback.OLD_COMBAT;
 
         public ClassicConfig() {
             super("classic_combat");
@@ -102,6 +106,11 @@ public class ServerConfig extends AbstractConfig {
     public static class BlockingConfig extends AbstractConfig {
         @Config(name = "allow_blocking", description = "Allow blocking with swords, which will reduce most incoming attacks by 50% and render a parry animation.")
         public boolean allowBlocking = true;
+        @Config(name = "prioritize_shield", description = "Prioritize shield blocking over sword blocking in case both items are held at the same time.")
+        public boolean prioritizeShield = true;
+        @Config(name = "knockback_reduction", description = "Percentage to reduce knockback by while sword blocking.")
+        @Config.DoubleRange(min = 0.0, max = 1.0)
+        public double knockbackReduction = 0.25;
 
         public BlockingConfig() {
             super("sword_blocking");
@@ -109,10 +118,12 @@ public class ServerConfig extends AbstractConfig {
     }
 
     public static class AdjustmentsConfig extends AbstractConfig {
-        @Config(name = "require_sweeping_edge", description = "Is the sweeping edge enchantment required to perform a sweep attack.")
+        @Config(category = "sweeping", name = "require_sweeping_edge", description = "Is the sweeping edge enchantment required to perform a sweep attack.")
         public boolean sweepingRequired = true;
-        @Config(name = "prioritize_shield", description = "Prioritize shield blocking over sword blocking in case both items are held at the same time.")
-        public boolean prioritizeShield = true;
+        @Config(category = "sweeping", name = "half_sweeping_damage", description = "Only apply half the sweeping damage to indirectly hit mobs as seen in most recent combat test snapshots.")
+        public boolean halfSweepingDamage = true;
+        @Config(category = "sweeping", name = "no_sweeping_when_sneaking", description = "Do not perform sweep attacks when sneaking.")
+        public boolean noSneakSweeping = false;
         @Config(name = "sprint_attacks", description = "Attacking will no longer stop the player from sprinting. Very useful when swimming, so you can fight underwater without being stopped on every hit.")
         public boolean sprintAttacks = true;
         @Config(name = "canceled_attack_sounds", description = {"Prevent various attack sounds added for the cooldown mechanic from playing.", EntryCollectionBuilder.CONFIG_DESCRIPTION})
@@ -121,14 +132,15 @@ public class ServerConfig extends AbstractConfig {
         public boolean noDamageIndicators = true;
         @Config(name = "min_hitbox_size", description = {"Force all entity hitboxes to have a cubic size of at least 0.9 blocks.", "This only affects targeting an entity, no collisions or whatsoever. Useful for hitting e.g. bats, rabbits, silverfish, fish, and most baby animals."})
         public boolean minHitboxSize = true;
-        @Config(name = "half_sweeping_damage", description = "Only apply half the sweeping damage to indirectly hit mobs as seen in most recent combat test snapshots.")
-        public boolean halfSweepingDamage = true;
-        @Config(name = "no_sweeping_when_sneaking", description = "Do not perform sweep attacks when sneaking.")
-        public boolean noSneakSweeping = false;
         @Config(name = "hold_attack_button", description = "Holding down the attack button keeps attacking continuously. No more spam clicking required.")
         public boolean holdAttackButton = true;
         @Config(name = "swing_through_grass", description = "Hit mobs through blocks without a collision shape such as tall grass without having to break the block first.")
         public boolean swingThroughGrass = true;
+        @Config(name = "shield_knockback_reduction", description = "Percentage to reduce knockback by while blocking with a shield.")
+        @Config.DoubleRange(min = 0.0, max = 1.0)
+        public double knockbackReduction = 0.5;
+        @Config(name = "shield_knockback_fix", description = "Fix shields not knocking back attackers (see MC-147694).")
+        public boolean shieldKnockbackFix = true;
 
         public Set<SoundEvent> canceledAttackSounds;