+++ src/main/java/fuzs/goldenagecombat/mixin/LivingEntityMixin.java
@@ -6,10 +6,9 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
-import org.spongepowered.asm.mixin.injection.Slice;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.*;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin extends Entity {
@@ -23,4 +22,14 @@ public abstract class LivingEntityMixin extends Entity {
         if (!GoldenAgeCombat.CONFIG.server().classic.backwardsWalking) return oldRotation;
         return 0.0F;
     }
+
+    @Inject(method = "blockedByShield", at = @At("HEAD"), cancellable = true)
+    protected void blockedByShield$head(LivingEntity target, CallbackInfo callbackInfo) {
+        if (!GoldenAgeCombat.CONFIG.server().adjustments.shieldKnockbackFix) return;
+        this.knockback(0.5, target.getX() - this.getX(), target.getZ() - this.getZ());
+        callbackInfo.cancel();
+    }
+
+    @Shadow
+    public abstract void knockback(double p_147241_, double p_147242_, double p_147243_);
 }