+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog].
 ## [v3.1.0-1.18.1] - 2022-02-08
 ### Added
 - Added attack reach attribute from combat test snapshots, it is available for swords, hoes, and tridents
+- Made players blocking with a sword take slightly less knockback
+- Made armor attributes render just as they do on tools and weapons
 ### Changed
 - Moved item attribute changes to their own config section with new options
+- Also moved some more options around in the config
+### Fixed
+- Fixed the mod crashing servers due to a faulty import
 
 ## [v3.0.0-1.18.1] - 2022-02-07
 - Ported to Minecraft 1.18